backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.14.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
aiosqlite>=0.20.0,<1.0.0
bcrypt>=4.1.0,<5.0.0
